New website layout and theme
Now projects are a single block view with multiple tabs. Separated 'my tasks' and 'all tasks' views onto their own pages.
Main sidebar always displays the group page links and also has a dropdown for the individual project pages. Moved search
button to be with main content, because when it was on the right page edge it could obscure the scrollbar in some cases.
Fixed some bugs with the search query and added ui element when nothing was found. Added multiple icons and adjusted page
theme.