Fix bug with task state edit
By not selecting a new state the select-element would have the the current option-element selected, which is disabled.
Therefore the form would be missing a key when sent creating an error. The current state option should be disabled,
because it helps the user know which of the options was previously selected when looking at the dropdown.