Fix some bugs with settings editing
some special characters would be escaped by jinja so when applying the name to the editing text field when edit mode was opened,
the text would show the escaped character. current fix makes it work with most special characters. some special characters that
are common in html are still escaped, but not sure this could be fixed without creating an XSS vulnerability.